Assersion問題の解決。doubleをfloatに変換

+++ ar2016/include/moving/moving.h
@@ -6,34 +6,34 @@ class Moving {
 public:
 	Moving() {}
 
-	Moving(double time_step, const Eigen::Vector2d& accel, const Eigen::Vector2d& velocity)
+	Moving(float time_step, const Eigen::Vector2f& accel, const Eigen::Vector2f& velocity)
 		: m_time_step(time_step), m_accel(accel), m_velocity(velocity)
 	{}
 
 	void updatePosition(Eigen::Vector2i& position);
 
-	void setAccel(const Eigen::Vector2d& accel) {
+	void setAccel(const Eigen::Vector2f& accel) {
 		m_accel = accel;
 	}
 
-	void setVelocity(const Eigen::Vector2d& velocity) {
+	void setVelocity(const Eigen::Vector2f& velocity) {
 		m_velocity = velocity;
 	}
 
-	Eigen::Vector2d getAccel() const
+	Eigen::Vector2f getAccel() const
 	{
 		return m_accel;
 	}
 
-	Eigen::Vector2d getVelocity() const
+	Eigen::Vector2f getVelocity() const
 	{
 		return m_velocity;
 	}
 
 private:
-	double m_time_step = 0.1;
-	Eigen::Vector2d m_accel = Eigen::Vector2d::Zero();
-	Eigen::Vector2d m_velocity = Eigen::Vector2d::Zero(); 
+	float m_time_step = 0.1;
+	Eigen::Vector2f m_accel = Eigen::Vector2f::Zero();
+	Eigen::Vector2f m_velocity = Eigen::Vector2f::Zero(); 
 
 	void updateVelocity();
